Removed bin level customization
It is not possible to pass CDK Constructs in from bin. They must be done
in the context of the stack. Which means it is not possible to pass
constructs in from bin, the stack itself must be customized.

Therefore the FoldingOnEcsstackProps I added are not useful and have
been removed.